More code verification

+++ src/ise/gameoflife/agents/GeneticAgent.java
@@ -70,7 +70,7 @@ public class GeneticAgent extends TestPoliticalAgent
 						opponentStagStrategy++;
 					}
 				}
-				opponentStagStrategy /= opponentHuntingHistory.size();
+				opponentStagStrategy /= (double)opponentHuntingHistory.size();
 			}
 		}
